requirements.txt
pydub==0.25.1
python-dotenv==1.0.1
openai==1.51.0
google-generativeai==0.8.2